refactor into a component to be reused
