+++ site/src/pages/TemplateSettingsPage/TemplateSchedulePage/TemplateScheduleForm.tsx
@@ -36,12 +36,12 @@ import { ScheduleDialog } from "./ScheduleDialog";
 import {
   AutostopRequirementDaysHelperText,
   AutostopRequirementWeeksHelperText,
-  AutostartRequirementDaysHelperText,
   convertAutostopRequirementDaysValue,
 } from "./AutostopRequirementHelperText";
 import { useTheme } from "@emotion/react";
 import Button from "@mui/material/Button";
 import FormHelperText from "@mui/material/FormHelperText";
+import { TemplateScheduleAutostart } from "components/TemplateScheduleAutostart/TemplateScheduleAutostart";
 
 const MS_HOUR_CONVERSION = 3600000;
 const MS_DAY_CONVERSION = 86400000;
@@ -441,82 +441,22 @@ export const TemplateScheduleForm: FC<TemplateScheduleForm> = ({
               </strong>
             </Stack>
           </Stack>
-
           {allowAdvancedScheduling && (
-            <Stack
-              direction="column"
-              width="100%"
-              alignItems="center"
-              css={{
-                marginBottom: "20px",
+            <TemplateScheduleAutostart
+              allow_user_autostart={form.values.allow_user_autostart}
+              autostart_requirement_days_of_week={
+                form.values.autostart_requirement_days_of_week
+              }
+              isSubmitting={isSubmitting}
+              onChange={async (
+                newDaysOfWeek: TemplateAutostartRequirementDaysValue[],
+              ) => {
+                await form.setFieldValue(
+                  "autostart_requirement_days_of_week",
+                  newDaysOfWeek,
+                );
               }}
-            >
-              <Stack
-                direction="row"
-                css={styles.ttlFields}
-                spacing={0}
-                alignItems="baseline"
-                justifyContent="center"
-              >
-                {(
-                  [
-                    { value: "monday", key: "Mon" },
-                    { value: "tuesday", key: "Tue" },
-                    { value: "wednesday", key: "Wed" },
-                    { value: "thursday", key: "Thu" },
-                    { value: "friday", key: "Fri" },
-                    { value: "saturday", key: "Sat" },
-                    { value: "sunday", key: "Sun" },
-                  ] as {
-                    value: TemplateAutostartRequirementDaysValue;
-                    key: string;
-                  }[]
-                ).map((day) => (
-                  <Button
-                    key={day.key}
-                    css={styles.dayButtons}
-                    // TODO: Adding a background color would also help
-                    color={
-                      form.values.autostart_requirement_days_of_week.includes(
-                        day.value,
-                      )
-                        ? "primary"
-                        : "secondary"
-                    }
-                    disabled={isSubmitting || !form.values.allow_user_autostart}
-                    onClick={async () => {
-                      if (
-                        !form.values.autostart_requirement_days_of_week.includes(
-                          day.value,
-                        )
-                      ) {
-                        await form.setFieldValue(
-                          "autostart_requirement_days_of_week",
-                          form.values.autostart_requirement_days_of_week.concat(
-                            day.value,
-                          ),
-                        );
-                      } else {
-                        await form.setFieldValue(
-                          "autostart_requirement_days_of_week",
-                          form.values.autostart_requirement_days_of_week.filter(
-                            (obj) => obj !== day.value,
-                          ),
-                        );
-                      }
-                    }}
-                  >
-                    {day.key}
-                  </Button>
-                ))}
-              </Stack>
-              <FormHelperText>
-                <AutostartRequirementDaysHelperText
-                  allowed={form.values.allow_user_autostart}
-                  days={form.values.autostart_requirement_days_of_week}
-                />
-              </FormHelperText>
-            </Stack>
+            />
           )}
 
           <Stack direction="row" alignItems="center">
